requirements.txt
DrissionPage==4.1.0.9
colorama==0.4.6
python-dotenv
pyinstaller
requests